[refactor] 폴더 내부 파일명 변경
index.js 와 style.js 로 통일
파일경로를 간단히 하기 위해 Menu.js 대신 index.js 사용
스타일 관련 코드 분리 위해 style.js 생성
