+++ sidebar/src/components/UI/atoms/abandon_button/index.js
@@ -1,17 +1,9 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import DeleteIcon from '@material-ui/icons/Delete';
 import Button from '@material-ui/core/Button';
+import useStyles from './style';
 
-const useStyles = makeStyles((theme) => ({
-  button: {
-    margin: theme.spacing(1),
-    color: '#616464',
-    borderColor: '#e0e0e0',
-  },
-}));
-
-const AbandonButton = ({ handleClickOpen }) => {
+const AbandonButton = ({ handleOpen }) => {
   const classes = useStyles();
 
   return (
@@ -21,7 +13,7 @@ const AbandonButton = ({ handleClickOpen }) => {
         className={classes.button}
         startIcon={<DeleteIcon />}
         size="small"
-        onClick={handleClickOpen}
+        onClick={handleOpen}
       >
         포기
       </Button>
