+++ sidebar/src/components/UI/atoms/abandon_button/style.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  button: {
+    margin: theme.spacing(1),
+    color: '#616464',
+    borderColor: '#e0e0e0',
+  },
+}));
+
+export default useStyles;
