+++ sidebar/src/components/UI/atoms/go_back_button/style.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    marginRight: theme.spacing(2),
+    width: '48px',
+    marginLeft: '-12px',
+  },
+}));
+
+export default useStyles;
