+++ sidebar/src/components/UI/atoms/text/style.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import styled from 'styled-components';
 
@@ -13,14 +12,4 @@ const TextWrapper = styled.div`
   font-size: ${(props) => props.fontSize};
 `;
 
-const text = ({ value, fontSize }) => {
-  const classes = useStyles();
-
-  return (
-    <TextWrapper className={classes.root} fontSize={fontSize}>
-      {value}
-    </TextWrapper>
-  );
-};
-
-export default text;
+export { useStyles, TextWrapper };
