+++ sidebar/src/components/UI/organisms/abandon_dialog/style.js
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  dialogContent: {
+    width: '400px',
+  },
+}));
+
+export default useStyles;
