+++ sidebar/src/components/UI/organisms/menu_drawer/style.js
@@ -1,0 +1,34 @@
+import { makeStyles } from '@material-ui/core/styles';
+import { deepOrange } from '@material-ui/core/colors';
+
+const drawerWidth = 400;
+
+const useStyles = makeStyles((theme) => ({
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+  },
+  drawerPaper: {
+    width: drawerWidth,
+  },
+  drawerHeader: {
+    display: 'flex',
+    alignItems: 'center',
+    padding: theme.spacing(0, 1),
+    ...theme.mixins.toolbar,
+    justifyContent: 'flex-end',
+  },
+  list: {
+    paddingLeft: '10px',
+  },
+  listIcon: {
+    minWidth: '45px',
+  },
+  orange: {
+    color: theme.palette.getContrastText(deepOrange[500]),
+    backgroundColor: deepOrange[500],
+    marginRight: '10px',
+  },
+}));
+
+export default useStyles;
