--- sidebar/src/components/templates/menu/Menu.js
@@ -1,147 +1,0 @@
-import React, { useState } from 'react';
-import clsx from 'clsx';
-import { makeStyles, useTheme } from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
-import Typography from '@material-ui/core/Typography';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import ChevronRightIcon from '@material-ui/icons/ChevronRight';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import ListAltIcon from '@material-ui/icons/ListAlt';
-import PeopleIcon from '@material-ui/icons/People';
-import EqualizerIcon from '@material-ui/icons/Equalizer';
-import FeedbackIcon from '@material-ui/icons/Feedback';
-import SettingsIcon from '@material-ui/icons/Settings';
-import ExitToAppIcon from '@material-ui/icons/ExitToApp';
-import Avatar from '@material-ui/core/Avatar';
-import { deepOrange } from '@material-ui/core/colors';
-import { Link } from 'react-router-dom';
-
-const drawerWidth = 400;
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-  },
-  appBar: {
-    backgroundColor: '#333333',
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  hide: {
-    display: 'none',
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-  },
-  drawerPaper: {
-    width: drawerWidth,
-  },
-  drawerHeader: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-    justifyContent: 'flex-end',
-  },
-  list: {
-    paddingLeft: '10px',
-  },
-  listIcon: {
-    minWidth: '45px',
-  },
-  orange: {
-    color: theme.palette.getContrastText(deepOrange[500]),
-    backgroundColor: deepOrange[500],
-    marginRight: '10px',
-  },
-}));
-
-const Menu = ({ title }) => {
-  const classes = useStyles();
-  const theme = useTheme();
-  const [open, setOpen] = useState(false);
-  const menuItems = [
-    { title: '내 목표', icon: <ListAltIcon />, path: '/' },
-    { title: '팔로우', icon: <PeopleIcon />, path: '/follow' },
-    { title: '통계', icon: <EqualizerIcon />, path: '/statistics' },
-    { title: '피드', icon: <FeedbackIcon />, path: '/feed' },
-    { title: '설정', icon: <SettingsIcon />, path: '/settings' },
-    { title: '로그아웃', icon: <ExitToAppIcon />, path: '/logout' },
-  ];
-
-  const toggleDrawer = (open) => setOpen(open);
-
-  return (
-    <div className={classes.root}>
-      <CssBaseline />
-      <AppBar position="fixed" className={clsx(classes.appBar)}>
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={() => toggleDrawer(true)}
-            edge="start"
-            className={clsx(classes.menuButton, open && classes.hide)}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" noWrap>
-            {title}
-          </Typography>
-        </Toolbar>
-      </AppBar>
-      <Drawer
-        className={classes.drawer}
-        anchor="left"
-        open={open}
-        onClose={() => toggleDrawer(false)}
-        classes={{
-          paper: classes.drawerPaper,
-        }}
-      >
-        <div role="presentation" onClick={() => toggleDrawer(false)}>
-          <div className={classes.drawerHeader}>
-            <IconButton onClick={() => toggleDrawer(false)}>
-              {theme.direction === 'ltr' ? <ChevronLeftIcon /> : <ChevronRightIcon />}
-            </IconButton>
-          </div>
-          <Divider />
-          <List className={classes.list}>
-            <ListItem button key="사용자 정보">
-              <Avatar className={classes.orange}>H</Avatar>
-              <ListItemText primary="사용자 정보" />
-            </ListItem>
-          </List>
-          <Divider />
-          <List className={classes.list}>
-            {menuItems.map((item, index) => (
-              <Link to={item.path} style={{ textDecoration: 'none' }} key={index}>
-                <ListItem button key={index}>
-                  <ListItemIcon className={classes.listIcon}>{item.icon}</ListItemIcon>
-                  <ListItemText primary={item.title} style={{ color: '#000000' }} />
-                </ListItem>
-              </Link>
-            ))}
-          </List>
-        </div>
-      </Drawer>
-    </div>
-  );
-};
-
-export default Menu;
